+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This file documents recent notable changes to this project. The format of this
 file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Bump bincode crate to 2.0 and modified the related code.
+
 ## [0.28.0] - 2025-09-30
 
 ### Added
@@ -1099,6 +1105,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - An initial version.
 
+[Unreleased]: https://github.com/aicers/review-web/compare/0.28.0...main
 [0.28.0]: https://github.com/aicers/review-web/compare/0.27.0...0.28.0
 [0.27.0]: https://github.com/aicers/review-web/compare/0.26.0...0.27.0
 [0.26.0]: https://github.com/aicers/review-web/compare/0.25.0...0.26.0